refactor for consistency
