+++ projects/wvr-elements/src/lib/shared/wvr-animation.component.ts
